requirements.txt
streamlit>=1.40.0
mysql-connector-python>=9.0.0
